Revert "הודעת ספירת העומר - לפי השקיעה"
This reverts commit acc12c787fbd06122a803868747201b1c9376d0b.
